requirements.txt
streamlit
streamlit-webrtc
openai-whisper
av
requests
python-dotenv